OP-440: Remove unnecessary traits and move methods to existing traits
